requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
alembic
pydantic
pydantic-settings
stripe>=7.0
httpx>=0.27
szamlazz.py>=1.5
